backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
pydantic>=2.10.0
pydantic-settings>=2.7.0
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
sse-starlette==1.8.2
httpx==0.24.1
aiohttp==3.9.1
openai==1.3.7
replicate==0.15.4
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-html==4.1.1
pytest-cov==7.0.0
factory-boy==3.3.0
respx==0.20.2
Pillow==11.3.0